refactor(worker-javascript): remove dead Core.js wrappers and small cleanups
Drop 11 Core.js delegating wrappers verified to have no callers outside
their owning modules (grepped across all package source + test directories
and confirmed none are exposed via `coreFunctions`):

- assertDiagnosticIsValid, findShadowedChildInSerializedComponents,
  expandShadowingComposite, addUndisplayableErrorChildrenToAncestor,
  markWithheldReplacementsInactive, findChildGroupNoAdapters,
  modifyStateDefsToBeShadows, modifyStateDefToDeleteVariableReferences,
  processChildChangesAndRecurseToShadows, calculateEssentialVariableChanges,
  calculatePrimitiveChildChanges

Move `numberAnswers` from the bottom of Core.js to `utils/answer.js`
(its sole caller is `generateDast`; pure tree-walking helper, no Core
dependency).

Small cleanups surfaced by the review:
- EssentialValueWriter: drop redundant `!compStateObj.hasEssential` check
  inside the non-array branch (already handled by the early-return one
  loop iteration earlier)
- StalenessPropagator: drop unused `arrayStateVariable` alias (byte-identical
  to `arrayVariableName` 13 lines above)
- StateVariableEvaluator: simplify `delete this.core._components[component
  .componentIdx].X` to `delete component.X` (caller passes the same object)
- EssentialValueWriter: convert nested `convert_md_array` arrow-const to a
  function declaration per AGENTS.md (not reassigned)
- Drop unused `deepClone` imports in CompositeReplacementUpdater and
  StalenessPropagator (sole references were commented-out console.logs)
- Fix typos: "intrepret" x4 in StalenessPropagator;
  "additionalcompositesWithAdjustedReplacements" -> camelCase in
  CompositeReplacementUpdater

Net -83 lines. Build passes; no new TS warnings introduced; bundle size
drops 2KB.
