+++ packages/doenetml-worker-javascript/src/EssentialValueWriter.ts
@@ -344,17 +344,6 @@ export class EssentialValueWriter {
                     }
                 } else {
                     // don't have array
-
-                    if (!compStateObj.hasEssential) {
-                        this.core.addDiagnostic({
-                            type: "info",
-                            message: `can't update state variable ${vName} of component ${cIdx}, as it does not have an essential state variable.`,
-                            position: this.core._components[cIdx].position,
-                            sourceDoc: this.core._components[cIdx].sourceDoc,
-                        });
-                        continue;
-                    }
-
                     if (compStateObj.set) {
                         comp.essentialState[essentialVarName] =
                             compStateObj.set(newComponentStateVariables[vName]);
@@ -810,28 +799,26 @@ export class EssentialValueWriter {
                                     // to an object with multidimesional arrayKeys
                                     // where each array key is a concatenation of the array indices, joined by commas
 
-                                    let convert_md_array = (
+                                    function convert_md_array(
                                         array: any,
                                         n_dim: number,
-                                    ): Record<string, any> => {
+                                    ): Record<string, any> {
                                         if (n_dim === 1) {
                                             return Object.assign({}, array);
-                                        } else {
-                                            let new_obj: Record<string, any> =
-                                                {};
-                                            for (let ind in array) {
-                                                let sub_obj = convert_md_array(
-                                                    array[ind],
-                                                    n_dim - 1,
-                                                );
-                                                for (let key in sub_obj) {
-                                                    new_obj[`${ind},${key}`] =
-                                                        sub_obj[key];
-                                                }
-                                            }
-                                            return new_obj;
                                         }
-                                    };
+                                        let new_obj: Record<string, any> = {};
+                                        for (let ind in array) {
+                                            let sub_obj = convert_md_array(
+                                                array[ind],
+                                                n_dim - 1,
+                                            );
+                                            for (let key in sub_obj) {
+                                                new_obj[`${ind},${key}`] =
+                                                    sub_obj[key];
+                                            }
+                                        }
+                                        return new_obj;
+                                    }
                                     Object.assign(
                                         arrayInstructionInProgress.desiredValue,
                                         convert_md_array(
